fix warnings
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51250 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ retex/renderer-base/src/main/java/com/himamis/retex/renderer/share/MonoScaleAtom.java
@@ -53,7 +53,7 @@ public class MonoScaleAtom extends ScaleAtom {
 	private double factor;
 
 	public MonoScaleAtom(Atom base, double factor) {
-		super(base, (double) factor, (double) factor);
+		super(base, factor, factor);
 		this.factor = factor;
 	}
